pyproject.toml
[tool.poetry]
name = "homework 6"
version = "0.1.0"
description = ""
authors = ["frutez <glebka.yatsenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
djoser = "^2.2.2"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
